Add files via upload
The machine learning model is trained using the test.csv dataset  and predictions are done on the test.csv dataset. The predicted results stored in the submission file and the code is stored in the Titanic_cpy_notebook.ipynb file. Here 0 means the person died and 1 means the person survived.
The column to be predicted is the Survived column in the train set. 

Dataset Desc.

PassengerId:- The id number of passenger.
Survived:- The Passenger survived or not.
PClass:- The class in which the passenger was travelling.
Name:- Name of the passenger
Sex :- Gender of the passenger
Age:- Age of the passenger.
SibSp:- Number of siblings travelling with the passenger
Parch :- Number of parents travelling with the passenger.
Ticket:- The Ticker number of the passenger.
Fare:- fare paid by the passenger to travel.
Cabin:- The cabin number in which the passenger was travelling.
Embarked:- The port to which the passenger was travelling to.